requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
